+++ src/firewallfabrik/platforms/iptables/_policy_compiler.py
@@ -67,6 +67,7 @@ from firewallfabrik.compiler.processors._service import (
     SeparateTCPWithFlags,
     SeparateUserServices,
     VerifyCustomServices,
+    VerifyIcmpTypes,
     VerifyPortRanges,
 )
 from firewallfabrik.core.objects import (
@@ -534,6 +535,7 @@ class PolicyCompiler_ipt(PolicyCompiler):
         self.add(SeparateTCPWithFlags('split on TCP services with flags'))
         self.add(VerifyCustomServices('verify custom services'))
         self.add(VerifyPortRanges('verify port ranges'))
+        self.add(VerifyIcmpTypes('verify ICMP types'))
         self.add(VerifyAddressRanges('verify address ranges'))
         self.add(VerifyAddresses('verify addresses'))
         self.add(VerifyMacAddresses('verify MAC addresses'))