+++ src/firewallfabrik/platforms/nftables/_policy_compiler.py
@@ -64,6 +64,7 @@ from firewallfabrik.compiler.processors._policy import (
 )
 from firewallfabrik.compiler.processors._service import (
     SeparateTCPWithFlags,
+    VerifyIcmpTypes,
     VerifyPortRanges,
 )
 from firewallfabrik.core.objects import (
@@ -412,6 +413,7 @@ class PolicyCompiler_nft(PolicyCompiler):
         self.add(SeparateTCPWithFlags('split on TCP services with flags'))
         self.add(VerifyCustomServices('verify custom services'))
         self.add(VerifyPortRanges('verify port ranges'))
+        self.add(VerifyIcmpTypes('verify ICMP types'))
         self.add(VerifyAddressRanges('verify address ranges'))
         self.add(VerifyAddresses('verify addresses'))
         self.add(VerifyMacAddresses('verify MAC addresses'))